+++ cli/src/conf/conf_get.rs
@@ -41,7 +41,6 @@ mod tests {
         let sui_conf = SuiConf {
             pk: Some("123".to_string().into()),
             rpc_url: Some(reqwest::Url::parse("https://mainnet.sui.io").unwrap()),
-            gql_url: Some(reqwest::Url::parse("https://mainnet.sui.io/graphql").unwrap()),
         };
 
         let tools = HashMap::new();