+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [`0.7.0`] - Unreleased
 
+### `nexus-cli`
+
+#### Removed
+
+- `sui_gql_url` config field
+
 ### `nexus-sdk`
 
 #### Changed
 
 - transaction templates adjusted to allow for locking and finalizing gas payments
 - `GasSettlementUpdate` event replaced with `GasLockUpdate`
+- removed graphql client and changed event fetching to poll GRPC
 
 ## [`0.6.0`] - 2026-02-24